Angles in Samples
This code calculates all possible angles between all all cRCC in a sample. To use:
python3 c4_angleNhistoCalculator.py
A request will appear, that must be answered with, by example: d9NoLat,Synthetic. That is to say, with the conditions to generate parallels and if the samples are random real or synthetic.

To graph the histograms, just run graphHistogram.py